Move authentication-service onto loco-rs 1.0.1, the pilot for a family-wide migration
loco-rs 1.0.0/1.0.1 is the framework's first stable release (2026-07-29 /
2026-07-31), a large intentionally-breaking cut off 0.16.4: sea-orm 2.0,
sqlx 0.9, an AppContext builder pattern, feature renames, and 64-bit
primary keys, among others. authentication-service is documented as the
family's reference loco.rs crate, so it goes first — the recipe proven
here is what the other 16 loco-rs crates in the monorepo get next.

What actually broke, and why:

- sea-orm's ConnectionTrait now splits `execute`/`query_one` (typed
  StatementBuilder, `&stmt`) from `execute_raw`/`query_one_raw` (a plain
  Statement, owned). The rate limiter's advisory-lock/window queries and
  the magic-link atomic-consume statement are genuinely raw SQL, so they
  move to the `_raw` methods rather than fighting the new trait bound.
- sea-query 1.0 moved `.eq()` off an inherent Expr method onto the new
  `ExprTrait`, which has to be imported or the call silently resolves to
  `core::cmp::PartialEq::eq` instead (a `bool`, not a `SimpleExpr` — the
  compiler catches this, but the error is easy to misread).
- loco's `ColType::PkAuto` now generates a 64-bit primary key
  (`big_pk_auto`/BIGINT) instead of 32-bit. The three generated entities
  (users, sessions, auth_events) and the one hand-written struct field
  that carries a row id (`audit_integrity::Report::mismatched`) move
  from `i32` to `i64` to match what the migration now actually creates.
  `cargo loco db entities` did not regenerate these on its own here (ran
  clean, touched nothing) — regenerated by hand against a real Postgres
  `\d users` to confirm the column type first.
- `serde_yaml` → `serde_yaml_ng`'s stricter parser rejected a duplicate
  `mailer.smtp.host/port/secure` block in config/development.yaml that
  the old parser silently took last-wins on. Sat there fine until a
  stricter parser needed to check.

Everything else — AppContext construction (never built by hand, goes
through create_app), the DownloadWorker/Queue registration, the Error
non-exhaustive match, the fallback/remote-IP middleware defaults — is
already compatible or unused here, so no change was needed.

Verified: cargo build --all-targets, fmt, clippy -D warnings, cargo test
(84 passed) all clean; the DB-gated suite (38 tests) green against a
fresh Postgres 18 with the real migrated (bigint-PK) schema.

+++ authentication/authentication-service-with-loco/src/compliance/audit_integrity.rs
@@ -153,7 +153,7 @@ pub struct AuditIntegrityReport {
     /// Rows naming a key or scheme this service cannot check.
     pub mac_unverifiable: usize,
     /// Rows whose content did not match what was stored.
-    pub mismatched: Vec<i32>,
+    pub mismatched: Vec<i64>,
     /// `true` when no mismatch was found.
     pub verified: bool,
     /// What this result does and does not attest to, carried in the
